requirements.txt
click==6.6
coreapi==1.32.3
coreapi-cli==1.0.0
Django==1.9.8
django-crispy-forms==1.6.0
django-extensions==1.7.1
django-filter==0.14.0
django-rest-swagger==2.0.5
djangorestframework==3.4.0
itypes==1.1.0
openapi-codec==1.0.0
psycopg2==2.6.2
pytz==2016.6.1
requests==2.11.1
simplejson==3.8.2
six==1.10.0
uritemplate==0.6
